requirements.txt
Flask==3.0.3
Flask-Limiter==3.8.0
beautifulsoup4==4.12.3
requests==2.32.3
psycopg[binary]==3.2.3
pytz==2024.2
python-dotenv==1.0.1
